Add support for user-specific data.
Provides extra method in Java ZxcvbnHelper class which takes a list of strings
for data such as names or other biographical data related to the user which
should not be used in passwords. The list is used to create a user-specific
dictionary matcher which is added to the matcher list before searching.